+++ tests/core/test_scripts.py
@@ -1,12 +1,10 @@
 from __future__ import annotations
 
-import csv
 import datetime
 import json
-import os
 import random
 from io import StringIO
-from unittest.mock import MagicMock, PropertyMock, call, patch
+from unittest.mock import MagicMock, call, patch
 
 import pytest
 from freezegun import freeze_time
@@ -14,10 +12,6 @@ from sqlalchemy.exc import InvalidRequestError
 from sqlalchemy.orm import Session
 
 from api.lanes import create_default_lanes
-from api.overdrive import (
-    GenerateOverdriveAdvantageAccountList,
-    OverdriveAdvantageAccount,
-)
 from core.classifier import Classifier
 from core.config import CannotLoadConfiguration, Configuration, ConfigurationConstants
 from core.external_search import ExternalSearchIndex, Filter
@@ -2568,111 +2562,6 @@ class TestLoanNotificationsScript:
         ).value = ConfigurationConstants.FALSE
         self.script.do_run()
         assert self.script.process_loan.call_count == 0
-
-
-class TestGenerateOverdriveAdvantageAccountList:
-    def test_generate_od_advantage_account_list(self, db: DatabaseTransactionFixture):
-        output_file_path = "test-output.csv"
-        circ_manager_name = "circ_man_name"
-        parent_library_name = "Parent"
-        parent_od_library_id = "parent_id"
-        child1_library_name = "child1"
-        child1_advantage_library_id = "1"
-        child1_token = "token1"
-        child2_library_name = "child2"
-        child2_advantage_library_id = "2"
-        child2_token = "token2"
-        client_key = "ck"
-        client_secret = "cs"
-        library_token = "lt"
-
-        parent: Collection = db.collection(
-            name=parent_library_name,
-            protocol=ExternalIntegration.OVERDRIVE,
-            external_account_id=parent_od_library_id,
-        )
-        child1: Collection = db.collection(
-            name=child1_library_name,
-            protocol=ExternalIntegration.OVERDRIVE,
-            external_account_id=child1_advantage_library_id,
-        )
-        child1.parent = parent
-        overdrive_api = MagicMock()
-        overdrive_api.get_advantage_accounts.return_value = [
-            OverdriveAdvantageAccount(
-                parent_od_library_id,
-                child1_advantage_library_id,
-                child1_library_name,
-                child1_token,
-            ),
-            OverdriveAdvantageAccount(
-                parent_od_library_id,
-                child2_advantage_library_id,
-                child2_library_name,
-                child2_token,
-            ),
-        ]
-
-        overdrive_api.client_key.return_value = bytes(client_key, "utf-8")
-        overdrive_api.client_secret.return_value = bytes(client_secret, "utf-8")
-        type(overdrive_api).collection_token = PropertyMock(return_value=library_token)
-
-        with patch(
-            "core.scripts.GenerateOverdriveAdvantageAccountList._create_overdrive_api"
-        ) as create_od_api:
-            create_od_api.return_value = overdrive_api
-            GenerateOverdriveAdvantageAccountList(db.session).do_run(
-                cmd_args=[
-                    "--output-file-path",
-                    output_file_path,
-                    "--circulation-manager-name",
-                    circ_manager_name,
-                ]
-            )
-
-            with open(output_file_path, newline="") as csv_file:
-                csvreader = csv.reader(csv_file)
-                for index, row in enumerate(csvreader):
-                    if index == 0:
-                        assert "cm" == row[0]
-                        assert "collection" == row[1]
-                        assert "overdrive_library_id" == row[2]
-                        assert "client_key" == row[3]
-                        assert "client_secret" == row[4]
-                        assert "library_token" == row[5]
-                        assert "advantage_name" == row[6]
-                        assert "advantage_id" == row[7]
-                        assert "advantage_token" == row[8]
-                        assert "already_configured" == row[9]
-                    elif index == 1:
-                        assert circ_manager_name == row[0]
-                        assert parent_library_name == row[1]
-                        assert parent_od_library_id == row[2]
-                        assert client_key == row[3]
-                        assert client_secret == row[4]
-                        assert library_token == row[5]
-                        assert child1_library_name == row[6]
-                        assert child1_advantage_library_id == row[7]
-                        assert child1_token == row[8]
-                        assert "True" == row[9]
-                    else:
-                        assert circ_manager_name == row[0]
-                        assert parent_library_name == row[1]
-                        assert parent_od_library_id == row[2]
-                        assert client_key == row[3]
-                        assert client_secret == row[4]
-                        assert library_token == row[5]
-                        assert child2_library_name == row[6]
-                        assert child2_advantage_library_id == row[7]
-                        assert child2_token == row[8]
-                        assert "False" == row[9]
-                    last_index = index
-
-            os.remove(output_file_path)
-            assert last_index == 2
-            overdrive_api.client_key.assert_called_once()
-            overdrive_api.client_secret.assert_called_once()
-            overdrive_api.get_advantage_accounts.assert_called_once()
 
 
 class TestWorkConsolidationScript:
